move method because it caused circular import errors
