+++ osinfo/__init__.py
@@ -1,13 +1,15 @@
+import datetime
 import functools
 import logging
 
 import aiohttp.web
+import dateutil.parser
 import yaml
 
 import consts
 import eol
-import os_id_extension.model as osidmodel
 import osinfo.alpine
+import osinfo.model
 import osinfo.paths
 import util
 
@@ -37,7 +39,7 @@ def release_infos_from_cfg(
         return None
 
     return [
-        eol.os_release_info_from_release_cycle({
+        os_release_info_from_release_cycle({
             'eol': release_info_raw.get('eol_date'),
             'cycle': release_info_raw['name'],
             'latest': release_info_raw.get('greatest_version'),
@@ -73,7 +75,7 @@ def os_release_infos(
 
     if release_cycles:
         return [
-            eol.os_release_info_from_release_cycle(release_cycle)
+            os_release_info_from_release_cycle(release_cycle)
             for release_cycle in release_cycles
         ]
 
@@ -83,6 +85,41 @@ def os_release_infos(
     return release_infos_from_cfg(
         os_id=os_id,
         absent_ok=True,
+    )
+
+
+def os_release_info_from_release_cycle(
+    release_cycle: dict,
+) -> osinfo.model.OsReleaseInfo:
+    def eol_date() -> bool | datetime.datetime | None:
+        eol_date = release_cycle.get('extendedSupport')
+        if eol_date is None:
+            eol_date = release_cycle.get('eol')
+
+        # unfortunately, eol-api yields inconsistent values for `eol` attribute (bool vs timestamp)
+        if isinstance(eol_date, bool):
+            return eol_date
+        elif isinstance(eol_date, str):
+            return dateutil.parser.isoparse(eol_date)
+        else:
+            return None
+
+    def reached_eol(
+        eol_date: datetime.datetime | bool=eol_date(),
+    ) -> bool | None:
+        if isinstance(eol_date, bool):
+            return eol_date
+        elif isinstance(eol_date, datetime.datetime):
+            return eol_date < datetime.datetime.today()
+        else:
+            return None
+
+    return osinfo.model.OsReleaseInfo(
+        name=release_cycle['cycle'],
+        # not provided for all products
+        greatest_version=release_cycle.get('latest'),
+        eol_date=eol_date(),
+        reached_eol=reached_eol(),
     )
